+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - goget: no longer run `gometalinter --install`
 
+- golang, nodejs, rustup: modify PATH given by `env` command
+
 - rust: extract out separate "rustup" task
 
 - rust: extract out separate "rustc" task